Move AnalogInputCluster to __init__.py
